Extract request build helpers into a new class
